Fixed HeadpatScript, Merged art to Main Branch
Changes:
 - Kiria's graphics merged to main branch, not sure how it wasn't, but I'll figure it out later.
 - Edited properties of the status bars due to my own incompetence in setting them correctly.
 - Headpat Script now works properly, and updates the friendship values accordingly.
 - Back Button now added so you can leave the Headpat Minigame.

Known Bugs:
 - Stats do not yet decay over time to emulate neglect of friend.
 - Cleanup Minigame logic needs to be redone to actually fully randomize the pellet spot placement.
 - Oh dear god I tried making an animation object and it blew up my Unity project I think I did something wrong.
 - Food Minigame logic still nonexistent.
 - Food Minigame still nonexistent.
 - Programming deities angered at my cockiness, sacrificial goat required to appease them.

+++ Assets/Scripts/GameManager.cs
@@ -56,7 +56,7 @@ public class GameManager : MonoBehaviour
     {
         // Prevents us from having values above our set maximum 
         val = Mathf.Min(val, maxBarVal);
-        TheBar.sizeDelta = new Vector2((MaxBarLength / maxBarVal) * val, TheBar.rect.height);
+        TheBar.sizeDelta = new Vector2((float)(MaxBarLength / maxBarVal) * val, TheBar.rect.height);
     }
     //Generic version of UpdateBar that automatically updates all three rather than just one
     public void UpdateBar()
